Adds ability to edit project details
Each project now has an Edit button. When user clicks on the edit
button, a dialog opens up with the current details of the project. The
user can edit the input fields. If the user clicks Confirm, then
the existing input field values will both be used to update the values
in the projectList and on the screen.